need comments

+++ src/components/DeleteArticle/DeleteArticle.jsx
@@ -1,25 +1,30 @@
 import { deleteDoc, doc } from "firebase/firestore";
 import React from "react";
 import { db, storage } from "../../firebase-config";
-import {toast} from "react-toastify"
-import {ref, deleteObject} from "firebase/storage"
+import { toast } from "react-toastify";
+import { ref, deleteObject } from "firebase/storage";
 
 function DeleteArticle({ id, imageURL }) {
-  const handleDelete = async() => {
+  const handleDelete = async () => {
+    if (window.confirm("Art you sure you want to delete this article?")) {
       try {
         await deleteDoc(doc(db, "Articles", id));
-        toast("Article deleted successfully", {type: "success"})
-        const storageRef = ref(storage, imageURL)
-        await deleteObject(storageRef)
+        toast("Article deleted successfully", { type: "success" });
+        const storageRef = ref(storage, imageURL);
+        await deleteObject(storageRef);
       } catch (error) {
-        toast("Error deleting article", {type: "error"})
+        toast("Error deleting article", { type: "error" });
       }
+    }
   };
   return (
     <div>
-      <button className="btn btn-danger" onClick={handleDelete}>
-        Delete
-      </button>
+      <i
+        className="fa fa-times"
+        style={{ cursor: "pointer" }}
+        onClick={handleDelete}
+      >
+      </i>
     </div>
   );
 }
